# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "ProjectsController_listPublished",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectCardDto"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List published projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{slug}": {
      "get": {
        "operationId": "ProjectsController_getPublishedDetailBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailDto"
                }
              }
            }
          }
        },
        "summary": "Read a published project by slug",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "patch": {
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/publish": {
      "post": {
        "operationId": "ProjectsController_publishProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Publish a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/projects/{id}/unpublish": {
      "post": {
        "operationId": "ProjectsController_unpublishProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unpublish a project (admin)",
        "tags": [
          "projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "Xolvon Backend API",
    "description": "HTTP contract for the Xolvon backend (global prefix /api). Auth: `Authorization: Bearer <accessToken>` (JWT, issued by POST /api/auth/register and POST /api/auth/login). Public routes need no credentials; routes listed with bearer security reject missing or invalid tokens with 401.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token. Send it as `Authorization: Bearer <accessToken>`."
      }
    },
    "schemas": {
...
      "ProjectCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Project id (UUID v4)."
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "published"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "type",
          "summary",
          "status"
        ]
      },
...
      "ProjectDetailDto": {
        "type": "object",
        "properties": {}
      },
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Project title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 200 characters.",
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "description": "Controlled project type string, max 100 characters (not a final enum).",
            "maxLength": 100
          },
          "summary": {
            "type": "string",
            "description": "One-line summary, max 500 characters, sanitized on write.",
            "maxLength": 500
          },
          "problem": {
            "type": "string",
            "description": "Story section \"Problem\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          },
          "solution": {
            "type": "string",
            "description": "Story section \"Solution\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          },
          "techStack": {
            "type": "string",
            "description": "Raw tech-stack text (parsed at read boundary), max 1000 characters.",
            "maxLength": 1000
          },
          "result": {
            "type": "string",
            "description": "Story section \"Result\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          }
        },
        "required": [
          "title",
          "slug"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Project title, max 200 characters.",
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "description": "URL slug: lowercase letters, digits, hyphens. Max 200 characters.",
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "description": "Controlled project type string, max 100 characters (not a final enum).",
            "maxLength": 100
          },
          "summary": {
            "type": "string",
            "description": "One-line summary, max 500 characters, sanitized on write.",
            "maxLength": 500
          },
          "problem": {
            "type": "string",
            "description": "Story section \"Problem\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          },
          "solution": {
            "type": "string",
            "description": "Story section \"Solution\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          },
          "techStack": {
            "type": "string",
            "description": "Raw tech-stack text (parsed at read boundary), max 1000 characters.",
            "maxLength": 1000
          },
          "result": {
            "type": "string",
            "description": "Story section \"Result\", max 10000 characters, sanitized on write.",
            "maxLength": 10000
          }
        }
      },
...
    }
  }
}